ghost_uapi header restructuring.
Add a layer of indirection between the ghost_uapi.h and the actual ABI.

All source files depending on the ghost uapi include the "ghost_uapi.h"
header but the specific ABI is selected via the GHOST_SELECT_ABI macro
(not in this change).

An abi is entirely contained in the "/abi/<num>" directory.
For now we only expect this to include the "ghost.h" header directly copied
from the kernel source (include/uapi/linux/ghost.h). However it leaves the
door open for one-off, esoteric implementations that may need to re-implement
the core libraries (e.g. offload agent).

Note that the ABI directly affects the 'ghost' and 'agent' libraries. There
are two choices on how to deal with multiple ABIs in these libraries:

1. Snapshot the source files in the ABI directory so they are effectively
   readonly for the life of that ABI.

2. Judicious use of macros so that the same source files can be compiled
   with multiple ABI versions (e.g. #if GHOST_VERSION >= 75).

It would make sense for incremental changes to discriminate based on macros.

Note that we no longer export the actual uapi header via "ghost" or "base"
so any target that depends on the ABI must now explicitly take a dependency
on "ghost_uapi".

PiperOrigin-RevId: 512181913