Snapshot abi 84 and build the corresponding ghost libraries.
define_ghost_uapi(84) creates the 'ghost_uapi_84' target.

cc_library_ghost() with 'abi=84' creates the 'ghost_84' and
'agent_84' targets.

The agent has a dependency on the 'schedghostidle' bpf program so use
bpf_skel_ghost(schedghostidle) with 'abi=84' to create the skeleton
in an abi-specific directory (bpf/user/84/bpf_schedghostidle.skel.h).

Finally we need to include the abi-specific headers based on the
selected abi (GHOST_SELECT_ABI). This can probably be avoided if
we can use the preprocessor to generate the include file name.

PiperOrigin-RevId: 512200249