ghost: brho versus the verifier: array indexing
The array indexing fight has gone on long enough.  We were always at the mercy
of the compiler, which could always use different registers for the bounds
check and for the array index.  Depending on the size of the structs, the
inlining of msg handlers, and whatever, we'd run the risk of a small change in
one place leading to the compiler breaking the verification.

The only way around it is to write the code ourselves: inline BPF assembly,
both disgusting and glorious!

We could extend this to the arr_list and avl helpers, if needed.  That will
require some header surgery due to using different headers for BPF and for
userspace (which builds queue.bpf.h and avl.bpf.h for testing).  We'll see.

PiperOrigin-RevId: 513400782