ghost: brho versus the verifier: dereference of modified ctx ptr
The compiler can quite legally index into the context by adding to the pointer,
and then derefencing.  In the past, I tried convincing it not to with various
brittle inlining.  Until we can get upstream to allow that seemingly benign
operation (which might be never), copy the bpf_ghost_msg onto the stack.  That
should be the end of these errors, at least for BPF-MSG - which is the only
place I ever got them.

PiperOrigin-RevId: 513400852