ghost: no empty structs with C++
In C, compilers do the 'sane' thing and make an empty struct size 0.  That
might be either implementation-defined or undefined (in which case, the
compiler can just reformat your hard drive or post fake pictures to your
Instagram).  In C++, an empty struct has a size of 1.

We didn't notice with the idle_flux structs for a few reasons:
- userspace wasn't poking the flux_cpu fields.
- even if we did, idle_flux_cpu was at the end of flux_cpu, so the offsets of
  the other memebers were fine.
- the struct was aligned to 64, so that extra byte for idle_flux_cpu was just
  padding.

How did I notice?  I was working with a scheduler where the rest of the
flux_cpu struct was exactly 64 bytes.  The BPF compiler (clang -target bpf, C
code) thought struct idle_flux_cpu was 0 bytes, and the flux_cpu was exactly
64.  C++ userspace thought it was 128 bytes (65 aligned up) and was page
faulting when iterating across the cpus array.

The CHECKs in lib/flux.h might catch this if it happens again, but no
guarantee.  It'll at least catch the discrepancy between bpf and userspace.

PiperOrigin-RevId: 513824872