ghOSt CFS: Flag as const functions which do not change class state
When a class method does not affect the task state it's better annotated with `const`.
This allows for the usage of such method from functions which gets a `const`
reference to those objects, e.g. trace point functions.

Add `const` specifier to a few methods not affecting their objects state.

No functional changes.

PiperOrigin-RevId: 515324312