alwayslink=1 the ghost_uapi library
Defining an exported global symbol in the ghost_uapi library and then
forcing the linker to preserve all instances of the library code in the
final link step allows us to detect cases where an agent is taking a
dependency on more than one ABI (directly or indirectly via libraries).

Note that this is a coarse check in that it only validates that the agent
is linking against _one_ abi but it doesn't check whether it's the right
one.

PiperOrigin-RevId: 520374145