ghost: allow multiple flux sched instances to use the same cpu struct
You thought you saw the last of the macros in flux?!?!

Neither did I...

The comment explains it well - schedulers want their own per-cpu data, even if
they aren't currently allocated that cpu.  I thought I could get around it, but
even with biff, I already had a conflict.  One that might have come up in
practice too!  (I saw the bug and designed the fix, then saw it hit before
changing the code - a race on biff.current).

I've got a more complicated scheduler coming up that also wants sched-specific
per-cpu bits, even when we're not allocated that cpu.  So even if we solved
this for biff (which we could, in an ugly manner), we'd still have the problem
in the future.

On the bright side, the macros hide it pretty well.  Your sched's blob in
flux_cpu is just a function of the sched_id, which is easily accessible in
(most) all of the sched functions.

PiperOrigin-RevId: 520380137