ghost: add the prov flux scheduler
Prov is a scheduler-of-schedulers, like ROCI.  You configure it to provision a
certain number (and optionally a set) of cpus to give to the high prio child
scheduler, name "prio".  After that, cpus are given first to "next" and then to
"last".

Currently, it can only handle being a root scheduler (same as ROCI), but
eventually I'd like to be able to have both roci and prov be able to be child
schedulers.

PiperOrigin-RevId: 521483359